+++ src/features/Character/components/Skill/index.js
@@ -6,12 +6,30 @@ import get from 'lodash/get';
 import TooltipTrigger from 'common/components/TooltipTrigger';
 import Icon from 'common/components/Icon';
 
+import EmptySkill from './Empty';
 import styles from './styles.less';
 
-const Skill = ({ data, className }: { data?: {}, className?: string }) => (
-  <TooltipTrigger type="skill" data={data || 'No Skill Selected'}>
-    <Icon src={get(data, 'icon')} size="mediumSmall" className={cx(styles.skill, className)} />
-  </TooltipTrigger>
-);
+type Props = {
+  data?: {},
+  className?: string,
+};
+
+const Skill = ({ data, className }: Props) => {
+  const error = get(data, 'error');
+
+  return (
+    <TooltipTrigger type="skill" data={data || 'No Skill Selected'}>
+      {(error || !data)
+        ? <EmptySkill />
+        : (
+          <Icon
+            src={get(data, 'icon')}
+            size="mediumSmall"
+            className={cx(styles.skill, className)}
+          />
+        )}
+    </TooltipTrigger>
+  );
+};
 
 export default Skill;